ping: add --healthy-only flag
Add `--healthy-only` flag to `scion ping`. When the --healthy-only option is set, ping first determines healthy paths through probing and chooses amongst them.

GitOrigin-RevId: 954892aeafa993df8614285cad53cb01ff822fb3
