+++ go/pkg/app/path/BUILD.bazel
@@ -1,0 +1,30 @@
+load("//lint:go.bzl", "go_library", "go_test")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["path.go"],
+    importpath = "github.com/scionproto/scion/go/pkg/app/path",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//go/lib/addr:go_default_library",
+        "//go/lib/pathpol:go_default_library",
+        "//go/lib/sciond:go_default_library",
+        "//go/lib/serrors:go_default_library",
+        "//go/lib/snet:go_default_library",
+        "//go/lib/snet/addrutil:go_default_library",
+        "//go/pkg/pathprobe:go_default_library",
+        "@com_github_fatih_color//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["path_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "//go/lib/snet:go_default_library",
+        "//go/lib/snet/path:go_default_library",
+        "//go/lib/xtest:go_default_library",
+        "@com_github_stretchr_testify//assert:go_default_library",
+    ],
+)
