+++ go/pkg/showpaths/BUILD.bazel
@@ -15,7 +15,7 @@ go_library(
         "//go/lib/serrors:go_default_library",
         "//go/lib/snet:go_default_library",
         "//go/lib/snet/addrutil:go_default_library",
-        "//go/pkg/app:go_default_library",
+        "//go/pkg/app/path:go_default_library",
         "//go/pkg/pathprobe:go_default_library",
     ],
 )
